demo_repo/requirements.txt
pytest==7.4.0
requests==2.31.0
pyjwt==2.8.0
flask==2.3.2
